Build with GNUmakefile/GNUstep

+++ NumericTerm.h
@@ -25,7 +25,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 }
 
 - initNumeric: (int) value;
-- (NSInteger) intValue;
+- (int) intValue;
 - (BOOL) isEqual: value;
 - (void) printForDebugger: (NSOutputStream *) stream;
 - (void) dealloc;
